Module skill_rating::elo::systems

Convenience functions for various popular rating systems using elo.

Functions

fide_k
FIDE calculates their k_factor depending on how many games you have played, and what you elo rating is.

game_fide
Convenience function for a game played with FIDE k_factor.

game_icc
Convenience function for a game played with ICC k_factor.

game_uscf
Convenience function for a game played with USCF k_factor.

icc_k
The ICC uses a global k_factor.

uscf_k
The USCF uses three differnt k_factors depending on you rating.